model_training_python/requirements.txt
datasets
transformers
trl
huggingface_hub
torch
bitsandbytes
peft
python-dotenv